+++ experimental/packages/opentelemetry-exporter-metrics-otlp-http/test/metricsHelper.ts
@@ -27,6 +27,7 @@ import { Resource } from '@opentelemetry/resources';
 import * as assert from 'assert';
 import { InstrumentationScope, VERSION } from '@opentelemetry/core';
 import {
+  Aggregation,
   AggregationTemporality,
   ExplicitBucketHistogramAggregation,
   MeterProvider,
@@ -55,6 +56,10 @@ class TestMetricReader extends MetricReader {
 
   protected onShutdown(): Promise<void> {
     return Promise.resolve(undefined);
+  }
+
+  selectAggregation() {
+    return Aggregation.Default();
   }
 
   selectAggregationTemporality() {